pyproject.toml
[project]
name = "airflow-learning"
version = "0.1.0"
description = "Apache Airflow samples"
authors = [
    {name = "Jose Meira"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "apache-airflow"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
